pipeline/requirements.txt
supabase>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
trafilatura>=1.8.0
